Refactored the discouraged method analyzers to use a shared base.
